requirements.txt
colorama==0.4.6
crayons==0.4.0
flake8==7.1.1
mccabe==0.7.0
numpy==2.1.3
opencv-python==4.10.0.84
piexif==1.1.3
pillow==10.4.0
pycodestyle==2.12.1
pyflakes==3.2.0
stegano==0.11.4
